pyproject.toml
[project]
name = "power-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "modbus-tk>=1.1.5",
    "pyserial>=3.5",
]
